+++ source/mdcii/mdcii/CMakeLists.txt
@@ -64,7 +64,7 @@ set(headers
     ${include_path}/menu/fps.hpp
     ${include_path}/cod/cod_parser.hpp
     ${include_path}/cod/basegad_dat.hpp
-    ${include_path}/cod/haeuser.hpp
+    ${include_path}/cod/buildings.hpp
     ${include_path}/cod/host_gad.hpp
     ${include_path}/gam/chunk.hpp
     ${include_path}/gam/city.hpp
@@ -113,7 +113,7 @@ set(sources
     ${source_path}/menu/gamewindow.cpp
     ${source_path}/cod/cod_parser.cpp
     ${source_path}/cod/basegad_dat.cpp
-    ${source_path}/cod/haeuser.cpp
+    ${source_path}/cod/buildings.cpp
     ${source_path}/cod/host_gad.cpp
     ${source_path}/gam/chunk.cpp
     ${source_path}/gam/city.cpp
