+++ source/mdcii/mdcii/include/mdcii/menu/singleplayerwindow.hpp
@@ -34,8 +34,8 @@
 #include "bsh/zeireader.hpp"
 #include "bsh/zeitexture.hpp"
 #include "cod/basegad_dat.hpp"
+#include "cod/buildings.hpp"
 #include "cod/cod_parser.hpp"
-#include "cod/haeuser.hpp"
 #include "cod/host_gad.hpp"
 #include "files/files.hpp"
 #include "framebuffer/palette.hpp"
