+++ source/mdcii/mdcii/src/kamera.cpp
@@ -17,7 +17,7 @@
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 #include "kamera.hpp"
-#include "cod/haeuser.hpp"
+#include "cod/buildings.hpp"
 #include "files/files.hpp"
 #include <string>
